Test : 플레이어 게임오버시 GameOver 팝업 띄우기 구현중 백업

+++ Assets/Scripts/Spells/IceBallLv1.cs
@@ -72,7 +72,7 @@ public class IceBallLv1 : IceSpell
                 Player player = collision.gameObject.GetComponent<Player>();
                 if (player != null)
                 {
-                    player.GetHit(spellInfo.level);
+                    player.GetHit((sbyte)spellInfo.level);
                 }
                 else Debug.LogError("Player is null!");
             }
