+++ Assets/Scripts/Spells/WaterBallLv1.cs
@@ -69,7 +69,7 @@ public class WaterBallLv1 : WaterSpell
                 Player player = collision.gameObject.GetComponent<Player>();
                 if (player != null)
                 {
-                    player.GetHit(spellInfo.level);
+                    player.GetHit((sbyte)spellInfo.level);
                 }
                 else Debug.LogError("Player is null!");
             }
